Rename hook files
