+++ app/hooks/use-balance.ts
@@ -1,7 +1,7 @@
 import { ApolloClient } from "@apollo/client"
 import find from "lodash.find"
 import { getWallet } from "../graphql/query"
-import { useBTCPrice } from "./usePrice"
+import { useBTCPrice } from "./use-price"
 
 export const useUSDBalance = (client: ApolloClient<unknown>): number => {
   const btcPrice = useBTCPrice()
