+++ app/screens/debug-screen/debug-screen.tsx
@@ -14,7 +14,7 @@ import { walletIsActive } from "../../graphql/query"
 
 import type { ScreenType } from "../../types/jsx"
 import type { INetwork } from "../../types/network"
-import { useBTCPrice } from "../../hooks/usePrice"
+import { useBTCPrice } from "../../hooks/use-price"
 
 const styles = EStyleSheet.create({
   button: {
