+++ app/screens/move-money-screen/move-money-screen.tsx
@@ -25,7 +25,7 @@ import { LargeButton } from "../../components/large-button"
 import { Screen } from "../../components/screen"
 import { TransactionItem } from "../../components/transaction-item"
 import { balanceBtc, MAIN_QUERY, walletIsActive } from "../../graphql/query"
-import { useUSDBalance } from "../../hooks/useBalance"
+import { useUSDBalance } from "../../hooks/use-balance"
 import { translate } from "../../i18n"
 import { color } from "../../theme"
 import { palette } from "../../theme/palette"
